REFACTOR [assets, core, svg]
