+++ src/app/app.module.ts
@@ -41,15 +41,15 @@ import { NzToolTipModule } from 'ng-zorro-antd/tooltip';
 import { GithubOutline, RadarChartOutline, FileMarkdownOutline, DownloadOutline, MinusOutline, SaveOutline } from '@ant-design/icons-angular/icons';
 const icons = [GithubOutline, RadarChartOutline, FileMarkdownOutline, DownloadOutline, MinusOutline, SaveOutline];
 
-import { IssuesService } from './issues.service';
+import { IssuesService } from './services/issues.service';
 import {MarkdownModule} from 'ngx-markdown'
-import { Squiggly1Component } from './assets/squiggly1/squiggly1.component';
-import { Squiggly2Component } from './assets/squiggly2/squiggly2.component';
-import { CallbackComponent } from './callback/callback.component';
+import { Squiggly1Component } from './assets/svg/squiggly1.component';
+import { Squiggly2Component } from './assets/svg/squiggly2.component';
+import { CallbackComponent } from './assets/callback/callback.component';
 import { CleanupReportComponent } from './cleanup-report/cleanup-report.component';
 import { SharedReportComponent } from './shared-report/shared-report.component';
-import { HeaderComponent } from './header/header.component';
-import { LayoutComponent } from './layout/layout.component';
+import { HeaderComponent } from './assets/header/header.component';
+import { LayoutComponent } from './assets/layout/layout.component';
 
 @NgModule({
   declarations: [
