+++ src/app/assets/svg/squiggly1.component.ts
