+++ src/app/assets/svg/squiggly2.component.ts
