+++ src/app/core/interface.ts
